requirements.txt
numpy
pydicom
qtpy